requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
dataclasses-json==0.6.7
frozenlist==1.8.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpath-python==1.1.5
jsonpointer==3.1.1
langchain==1.2.15
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.0
langchain-openrouter==0.2.1
langchain-text-splitters==1.1.2
langgraph==1.1.8
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.32
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.4
openrouter==0.9.1
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
propcache==0.4.1
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
tenacity==9.1.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
wikipedia==1.4.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
